fix : DDayBox Custom hook 적용
